Split up channel.rs and refactor audio instance code
